requirements.txt
django>=3.2.0
djangorestframework>=3.12.0
yfinance>=0.1.63
pandas>=1.3.0
numpy>=1.19.0
scikit-learn>=0.24.0
python-dateutil>=2.8.0
requests>=2.25.0
django-cors-headers>=3.7.0 